+++ api/entities/drone.ts
