+++ api/ormconfig.json
@@ -7,9 +7,9 @@
 	"password": "paracord",
 	"synchronize": true,
 	"entities": [
-		"./drone.ts",
-		"./user.ts",
-		"./session.ts",
-		"./flight.ts"
+		"./entities/drone.ts",
+		"./entities/user.ts",
+		"./entities/session.ts",
+		"./entities/flight.ts"
 	]
 }
